+++ components/section/node.js
@@ -15,7 +15,7 @@ export default class SectionNode extends Component {
 
 	constructor(props) {
 		super(props);
-		this.state = { layout: props.getLayout, content: props.content, selectedSource: 'initial' };
+		this.state = { layout: props.layout, content: props.content, selectedSource: 'initial' };
 	}
 
 	loadContent() {
