+++ config/layouts/index.js
@@ -1,0 +1,23 @@
+import editorsPicksLayout from './editors-picks';
+import editorsPicksAdsLayout from './editors-picks-ads';
+import featuredSectionLayout from './featured-section';
+import mostPopularLayout from './most-popular';
+import mostPopularAdsLayout from './most-popular-ads';
+import myftLayout from './myft';
+import opinionLayout from './opinion';
+import topStoriesLayout from './top-stories';
+import topStoriesWithRelatedLayout from './top-stories-with-related';
+import VideoLayout from './video';
+
+export default {
+    'editors-picks': editorsPicksLayout,
+    'editors-picks-ads': editorsPicksAdsLayout,
+    'featured-section': featuredSectionLayout,
+    'most-popular': mostPopularLayout,
+    'most-popular-ads': mostPopularAdsLayout,
+    myft: myftLayout,
+    'opinion': opinionLayout,
+    'top-stories': topStoriesLayout,
+    'top-stories-with-related': topStoriesWithRelatedLayout,
+    video: VideoLayout
+}
