+++ config/pages.js
@@ -1,0 +1,20 @@
+import getSection from './pages';
+
+const pages = {
+	// ordered list of sections for the front page
+	'front-page': [
+		'top-stories',
+		'opinion',
+		'myft',
+		'editors-picks',
+		'most-popular',
+		'technology',
+		'markets',
+		'life-and-arts',
+		'videos'
+	]
+};
+
+export default (pageId, sectionsContent, flags) => {
+	return pages[pageId].map(sectionId => getSection(sectionId, sectionsContent[sectionId], flags));
+}
