+++ config/sections/life-and-arts.js
@@ -1,0 +1,18 @@
+export default () => ({
+    id: 'life-and-arts',
+    title: 'Life & Arts',
+    style: 'life-and-arts',
+    layoutId: 'featured-section',
+    size: {
+        default: 12,
+        M: 4
+    },
+    cols: {
+        meta: {
+            default: 12
+        },
+        content: {
+            default: 12
+        }
+    }
+})
