+++ config/sections/markets.js
@@ -1,0 +1,18 @@
+export default () => ({
+    id: 'markets',
+    title: 'Markets',
+    style: 'markets',
+    layoutId: 'featured-section',
+    size: {
+        default: 12,
+        M: 4
+    },
+    cols: {
+        meta: {
+            default: 12
+        },
+        content: {
+            default: 12
+        }
+    }
+})
