+++ config/sections/opinion.js
@@ -1,0 +1,17 @@
+export default () => ({
+    id: 'opinion',
+    title: 'Opinion',
+    style: 'opinion',
+    layoutId: 'opinion',
+    size: {
+        default: 12
+    },
+    cols: {
+        meta: {
+            default: 12
+        },
+        content: {
+            default: 12
+        }
+    }
+})
