+++ config/sections/technology.js
@@ -1,0 +1,18 @@
+export default () => ({
+    id: 'technology',
+    title: 'Technology',
+    style: 'technology',
+    layoutId: 'featured-section',
+    size: {
+        default: 12,
+        M: 4
+    },
+    cols: {
+        meta: {
+            default: 12
+        },
+        content: {
+            default: 12
+        }
+    }
+})
