+++ config/sections/videos.js
@@ -1,0 +1,18 @@
+export default () => ({
+    id: 'videos', // Note: calling this "video" breaks BB10
+    trackable: 'video',
+    title: 'Video',
+    style: 'video',
+    layoutId: 'video',
+    size: {
+        default: 12
+    },
+    cols: {
+        meta: {
+            default: 12
+        },
+        content: {
+            default: 12
+        }
+    }
+})
